+++ src/MyAccountingApp.Core/DTOs/MoneyResponse.cs
@@ -4,7 +4,7 @@ namespace MyAccountingApp.Core.DTOs;
 public class MoneyResponse
 {
     [JsonPropertyName("amount")]
-    public double Amount { get; init; }
+    public decimal Amount { get; init; }
 
     [JsonPropertyName("currency")]
     public string? Currency { get; init; }
